+++ tradingbot/lib/__init__.py
@@ -2,3 +2,4 @@ from . import time_ms
 from . import pandamex
 from . import dataset
 from . import line_notification
+from . import twitter_dataset
